Improved luminosity handling.

+++ Source/analysis.h
@@ -20,7 +20,7 @@
 
 class AnalysisZprime{
 public:
-  AnalysisZprime(const TString channel, const TString model, const int energy, const TString options, const int vegasIterations, const int vegasPoints, const double luminosity, const int btags, const bool discardComplex, const TString analysis_label);
+  AnalysisZprime(const TString channel, const TString model, const int energy, const TString options, const int vegasIterations, const int vegasPoints, const int luminosity, const int btags, const bool discardComplex, const TString analysis_label);
   virtual ~AnalysisZprime();
   TString GetOutputFilename();
 
@@ -84,7 +84,7 @@ private:
   TString m_options;
   int m_vegasIterations;
   int m_vegasPoints;
-  double m_luminosity;
+  int m_luminosity;
   const int m_btags;
   const bool m_discardComplex;
   const TString m_analysisLabel;
